Moved receiver update from ogn.collect.database to ogn.collect.receiver
