+++ ogn/collect/celery.py
@@ -26,7 +26,9 @@ def close_db(signal, sender):
 
 app = Celery('ogn.collect',
              include=["ogn.collect.database",
+                      "ogn.collect.heatmap",
                       "ogn.collect.logbook",
-                      "ogn.collect.heatmap"])
+                      "ogn.collect.receiver"
+                      ])
 
 app.config_from_envvar("OGN_CONFIG_MODULE")
